Added integration test to create a new task (#75)
* Added integration test for adding a new task

* Fixed path to script

* Remove integration test on android

* Remove brew update

* Added modify todo integration test
Updated add/modify screen for all projects except MVU to work with driver
Added timeouts for some driver commands in integration tests for CI
Fixed all builds for ios and android
Running integration tests on android in CI

* Updated add/modify screen for vanilla project to work with driver

* Added timeouts for all driver commands generated by add/edit tests
Added generator for .travis.yml, one job/project/platform

* Added generated .travis.yml

* Added .travis.yml generator

* Added timeouts to all driver commands
Fixed some failing integration tests

* Fixed exit code from running integration tests

* Using travis stages with yaml variables instead of generating
.travis.yml and removed dependency on travis for android
Added fastlane to vanilla app as placeholder for implementing deploy 🤓

* Added missing file

* Replaced Podfiles with latest autogenerated version

+++ scripts/runTests.sh
@@ -1,7 +1,28 @@
 #!/usr/bin/env bash
 
-# runs tests for a package that has tests
-# does not run integration tests
+show_help() {
+    printf "usage: $0 [--help] [--report] [<path to package>]
+
+Tool for running all unit and widget tests with code coverage.
+(run from root of repo)
+
+where:
+    <path to package>
+        run tests for package at path only
+        (otherwise runs all tests)
+    --report
+        run a coverage report
+        (requires lcov installed)
+    --help
+        print this message
+
+requires code_coverage package
+(install with 'pub global activate coverage')
+"
+    exit 1
+}
+
+# run unit and widget tests
 runTests () {
   cd $1;
   if [ -f "pubspec.yaml" ] && [ -d "test" ]; then
@@ -12,6 +33,9 @@ runTests () {
       flutter packages get
       flutter packages pub run build_runner build --delete-conflicting-outputs
     fi
+
+    escapedPath="$(echo $1 | sed 's/\//\\\//g')"
+
     # run tests with coverage
     if grep flutter pubspec.yaml > /dev/null; then
       echo "run flutter tests"
@@ -22,7 +46,6 @@ runTests () {
       fi
       if [ -d "coverage" ]; then
         # combine line coverage info from package tests to a common file
-        escapedPath="$(echo ${1:2} | sed 's/\//\\\//g')"
         sed "s/^SF:lib/SF:$escapedPath\/lib/g" coverage/lcov.info >> $2/lcov.info
         rm -rf "coverage"
       fi
@@ -30,26 +53,57 @@ runTests () {
       # pure dart
       echo "run dart tests"
       pub get
-      testFile="test/all_tests.dart"
       pub global run coverage:collect_coverage --port=8111 -o coverage.json --resume-isolates --wait-paused &
-      dart --pause-isolates-on-exit --enable-vm-service=8111 $testFile
+      dart --pause-isolates-on-exit --enable-vm-service=8111 "test/all_tests.dart"
       pub global run coverage:format_coverage --packages=.packages -i coverage.json --report-on lib --lcov --out lcov.info
       if [ -f "lcov.info" ]; then
         # combine line coverage info from package tests to a common file
-        escapedPath="$(echo ${1:2} | sed 's/\//\\\//g')"
         sed "s/^SF:.*lib/SF:$escapedPath\/lib/g" lcov.info >> $2/lcov.info
         rm lcov.info
       fi
       rm -f coverage.json
     fi
   fi
+  cd - > /dev/null
 }
 
-export -f runTests
+runReport() {
+    if [ -f "lcov.info" ] && ! [ "$TRAVIS" ]; then
+        genhtml lcov.info -o coverage --no-function-coverage -s -p `pwd`
+        open coverage/index.html
+    fi
+}
 
-# if running locally
-rm -f lcov.info
+if ! [ -d .git ]; then printf "\nError: not in root of repo"; show_help; fi
 
-# expects to find most packages at second directory level
-find . -maxdepth 2 -type d -exec bash -c 'runTests "$0" `pwd`' {} \;
-# find exits with 0
+case $1 in
+    --help)
+        show_help
+        ;;
+    --report)
+        if ! [ -z ${2+x} ]; then
+            printf "\nError: no extra parameters required: $2"
+            show_help
+        fi
+        runReport
+        ;;
+    *)
+        currentDir=`pwd`
+        # if no parameter passed
+        if [ -z $1 ]; then
+            rm -f lcov.info
+            dirs=(`find . -maxdepth 2 -type d`)
+            for dir in "${dirs[@]}"; do
+                runTests $dir $currentDir
+            done
+        else
+            if [[ -d "$1" ]]; then
+                runTests $1 $currentDir
+            else
+                printf "\nError: not a directory: $1"
+                show_help
+            fi
+        fi
+        ;;
+esac
+
